backend/diagnostics-service/requirements.txt
gradio
tensorflow-macos
tensorflow-metal
ultralytics
opencv-python
huggingface-hub
Pillow
flask